Introduce SamplingDtypeEnabler

See merge request ift/nifty!458

+++ nifty6/probing.py
@@ -100,9 +100,9 @@ def probe_with_posterior_samples(op, post_op, nprobes, dtype):
     sc = StatCalculator()
     for i in range(nprobes):
         if post_op is None:
-            sc.add(op.draw_sample(dtype=dtype, from_inverse=True))
+            sc.add(op.draw_sample(from_inverse=True))
         else:
-            sc.add(post_op(op.draw_sample(dtype=dtype, from_inverse=True)))
+            sc.add(post_op(op.draw_sample(from_inverse=True)))
 
     if nprobes == 1:
         return sc.mean, None
